Re-ported all vali changes to mesa3d

+++ src/gallium/auxiliary/util/u_debug.c
@@ -247,7 +247,7 @@ str_has_option(const char *str, const char *name)
        * we compare 'start' up to 'str-1' with 'name'. */
 
       while (1) {
-         if (!*str || !(isalnum(*str) || *str == '_')) {
+         if (!*str || !(isalnum((int)*str) || *str == '_')) {
             if (str-start == name_len &&
                 !memcmp(start, name, name_len)) {
                return TRUE;
